pyproject.toml
[project]
name = "friday-tony-stark-demo"
version = "0.1.0"
description = "Friday — A Tony Stark-inspired MCP server"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp",
    "httpx",
    "livekit-agents[deepgram,groq,openai,sarvam,silero]>=1.5.1",  # Deprecate later
    "livekit-plugins-google>=1.5.1",  # Deprecate
    "python-dotenv",
    "chromadb>=0.5.0",
    "sentence-transformers>=3.0.1",
    "pandas",
    "numpy",
    "scikit-learn",
    "schedule",
    "requests",
    "ollama",
    "speechrecognition",
    "pyttsx3",
    "streamlit",
    "faiss-cpu",
    "joblib",
]

[project.scripts]
friday = "server:main"
friday_voice = "agent_friday:dev"
autonomous = "friday.autonomous.run:main"
stark = "friday.main:main"  # New local main

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["server.py"]
